+++ src/frontend/apps/drive/src/features/ui/preview/preview.scss
@@ -1,12 +1,12 @@
-@use "./audio-player/audio-player.scss";
-@use "./files-preview/files-preview.scss";
-@use "./image-viewer/ImageViewer.scss";
-@use "./video-player/VideoPlayer.scss";
+@use "./FilesPreview.scss";
 @use "./components/duration-bar/DurationBar.scss";
 @use "./components/volume-bar/VolumeBar.scss";
 @use "./components/controls/PreviewControls.scss";
-@use "./not-supported/NotSupportedPreview.scss";
-@use "./suspicious/SuspiciousPreview.scss";
-@use "./wopi/WopiEditor.scss";
-@use "./error/ErrorPreview.scss";
-@use "./pdf-preview/PreviewPdf.scss";
+@use "./viewers/not-supported/NotSupportedPreview.scss";
+@use "./viewers/suspicious/SuspiciousPreview.scss";
+@use "./viewers/wopi/WopiEditor.scss";
+@use "./viewers/error/ErrorPreview.scss";
+@use "./viewers/audio-player/AudioPlayer.scss";
+@use "./viewers/image-viewer/ImageViewer.scss";
+@use "./viewers/video-player/VideoPlayer.scss";
+@use "./viewers/pdf-preview/PdfPreview.scss";
