+++ src/frontend/apps/drive/src/features/ui/preview/viewers/audio-player/AudioPlayer.scss
