+++ src/frontend/apps/drive/src/features/ui/preview/viewers/error/ErrorPreview.scss
