+++ src/frontend/apps/drive/src/features/ui/preview/viewers/image-viewer/ImageViewer.scss
@@ -6,14 +6,6 @@
   overflow: hidden;
   position: relative;
 
-  &__controls {
-    position: absolute;
-    bottom: 16px;
-    left: 50%;
-    transform: translateX(-50%);
-    z-index: 10;
-  }
-
   &__container {
     flex: 1;
     position: relative;
