+++ src/frontend/apps/drive/src/features/ui/preview/viewers/not-supported/NotSupportedPreview.scss
