+++ src/frontend/apps/drive/src/features/ui/preview/viewers/pdf-preview/OutdatedBrowserPreview.tsx
