+++ src/frontend/apps/drive/src/features/ui/preview/viewers/pdf-preview/pdfConsts.tsx
@@ -1,0 +1,5 @@
+export const TRANSITION_DELAY = 300;
+export const THUMBNAIL_HEIGHT = 178;
+export const THUMBNAIL_GAP = 12;
+export const ROW_HEIGHT = THUMBNAIL_HEIGHT + THUMBNAIL_GAP;
+export const OPEN_DELAY = 100;
