+++ src/frontend/apps/drive/src/features/ui/preview/viewers/suspicious/SuspiciousPreview.scss
