+++ src/frontend/apps/drive/src/features/ui/preview/viewers/video-player/VideoPlayer.scss
