+++ src/frontend/apps/drive/src/features/ui/preview/viewers/wopi/WopiEditor.scss
